made db.js an exportable module, sqlTest.js is now a file indicating how it works with sample query arguments displayed
